Separate out files to make bookmarklet easier
Sideloading extensions in Firefox appears to be not trivial because you
still need to sign things. This is even worse in mobile because it seems
you can't even do it as a temporary extension from the debug page. So
the bookmarklet remains useful till I get around to reading all about
the extension publication process.

Since the bookmarklet needs to also handle loading the scripts
dynamically and it gets triggered by the user (when the page has
presumably already loaded) it has slightly different code. So I've
separated out the core functionality into something that can get
included by both the bookmarklet and the extension. Then a little shell
script just cats them into a .gitignored file for easy copy-pasting into
a bookmarklet encoder tool.