Avoid permanently moving tiles above other elements
By default Draggable continuously increases the z-index of elements that
are dragged. The idea is if you're dragging things on top of each other,
you want the most recently dragged on top. But after every drag we reset
things back to the grid so we don't need that and it causes problems
because the tiles end up on top of e.g. the stats panel. Instead, we
bump up the z-index on drag and then reset it when the drag ends.