Prepare for publication of 0.4
Added the extension ID to the manifest, which is required to publish via
the API and also isn't supported by Chrome. Cool. The only effect seems
to be that Chrome complains and still runs the extension fine. Added an
AMO metadata file, which I guess I'll edit on version bumps to add the
release notes. Not super impressed with the ergonomics here...