Check in package.json for node
I'm still not checking in package-lock.json. To quote the comment I
added in the .gitignore:

I am making a principled decision not to check this into source control
in defiance of the node docs. I'm only using node for development tools
and I don't have any particular attachment to the current versions of
things, nor are any of the tools interacting with each other. I don't
particularly want to check in a many thousand line file that isn't
helpful for anything. I am, however, checking in the package.json, so
people can easily install the development tools.